Rename keep_alive to server. Worked on jump. Stable.

+++ cogs/system.py
@@ -1,5 +1,7 @@
+from cogs.utils.embed import send_error_embed
 from cogs.utils.time import now
 from discord.ext import commands
+from replit import db
 
 
 class System(commands.Cog):
@@ -15,6 +17,10 @@ class System(commands.Cog):
 
     @commands.command()
     async def jump(self, ctx):
+        if 'users' not in db.keys():
+            await send_error_embed(ctx, 'no_users')
+            return
+        
         await ctx.message.delete()
         await ctx.send(f'Pong: {round(self.bot.latency * 1000)}ms.')
